Refactor of GLM
