+++ brainbox/modeling/design_matrix.py
@@ -5,26 +5,62 @@ import scipy.sparse as sp
 import numba as nb
 
 
-class DesignMatrixConstructor:
+class DesignMatrix:
     """
     Design matrix constructor that will take in information about the temporal structure of a trial
     and allow the generation of a design matrix with specified regressors
     """
 
-    def __init__(self, trialsdf, vartypes, binwidth=0.02):
+    def __init__(self, trialsdf, vartypes=None, binwidth=0.02):
         """
-        I'll tell you more later, buddy
+        Class for generating design matrices to model neural data. Provides handy routines for
+        describing neural spiking activity using basis functions and other primitives. 
+
+        Based on work by Memming Park in:
+        Il Memming Park, Miriam LR Meister, Alex C Huk, & Jonathan W Pillow Nature Neuroscience 17,
+        1395-1403. (2014)
+        and the accompanying code in MATLAB.
+
+        Parameters
+        ----------
+        trialsdf : pandas.DataFrame
+            Dataframe in which each row is a trial, and each column is a trial-by-trial covariate.
+            This includes, optionally, continuous covariates like eye-position and wheel movement
+            per trial, which can be used with object-datatype dataframes. The length of vectors
+            stored in continuous-variable columns must match the length of the trial.
+
+            Obligatory columns for the dataframe are "trial_start" and "trial_end", which tell the
+            constructor which time points to associate with that trial.
+        vartypes : dict, optional
+            Dictionary of types for each of the columns in trialsdf. Columns must be of the types:
+            -- timing: timing events, in which the column values are times since the start of the
+                session of an event within that trial, e.g. stimulus onset.
+            -- value: scalars which describe a whole-trial value, such as stimulus contrast or
+                probability block
+            -- continuous: Columns which contain 1-D vectors per row that describe a covariate that
+                changes within the trial. e.g. pupil diameter.
+            Dictionary keys should be columns in trialsdf, values should be strings that are equal
+            to one of the above.
+
+            If vartypes is not passed, the constructor will assume you know what you are doing. Be
+            warned that this can result in the class failing in spectacular and vindictive ways.
+            by default None
+        binwidth : float, optional
+            Length of time bins which will be used for design matrix, by default 0.02
         """
         # Data checks #
-        if not all([name in vartypes for name in trialsdf.columns]):
-            raise KeyError("Some columns were not described in vartypes")
-        if not all([value in ('timing', 'continuous', 'value') for value in vartypes.values()]):
-            raise ValueError("Invalid values were passed in vartypes")
+        if vartypes is not None:
+            validtypes = ('timing', 'continuous', 'value')
+            if not all([name in vartypes for name in trialsdf.columns]):
+                raise KeyError("Some columns were not described in vartypes")
+            if not all([value in validtypes for value in vartypes.values()]):
+                raise ValueError("Invalid values were passed in vartypes")
 
         # Filter out cells which don't meet the criteria for minimum spiking, while doing trial
         # assignment
         self.vartypes = vartypes
-        self.vartypes['duration'] = 'value'
+        if vartypes is not None:
+            self.vartypes['duration'] = 'value'
         trialsdf = trialsdf.copy()  # Make sure we don't modify the original dataframe
         trbounds = trialsdf[['trial_start', 'trial_end']]  # Get the start/end of trials
         # Empty trial duration value to use later
@@ -47,12 +83,37 @@ class DesignMatrixConstructor:
         self.compiled = False
         return
 
+    def binf(self, t):
+        """
+        Function to bin time t into binwidth of DM
+
+        Parameters
+        ----------
+        t : float
+            time, in seconds
+
+        Returns
+        -------
+        int
+            number of time bins
+        """
+        return np.ceil(t / self.binwidth).astype(int)
+
     def add_covariate_timing(self, covlabel, eventname, bases,
                              offset=0, deltaval=None, cond=None, desc=''):
         """
-        Convenience wrapper for adding timing event regressors to the GLM. Automatically generates
-        a one-hot vector for each trial as the regressor and adds the appropriate data structure
-        to the model.
+        Convenience wrapper for adding timing event regressors to the design matrix.
+
+        Timing events are regressed against using basis functions, such as those generated by the
+        functions in this module, which are convolved with a kronecker delta at the time where the
+        event occurred. This operation is effectively a copy/paste of the basis functions, each of
+        which have their own column in the design matrix. This means that when we fit weights, a
+        single weight can govern the prediction of the model over a longer time period.
+
+        Can be offset, such that the bases functions are applied before or after the timing.
+
+        The height of the bases can be modified by a column in the design matrix, which will then
+        multiply the delta function which is convolved with the bases.
 
         Parameters
         ----------
@@ -66,8 +127,8 @@ class DesignMatrixConstructor:
             Each column in the array is used together to describe the response of a unit to that
             timing event.
         offset : float, seconds
-            Offset of bases functions relative to timing event. Negative values will ensure that
-
+            Offset of bases functions relative to timing event. Negative values mean bases will
+            be applied *before* the timing event by that amount.
         deltaval : None, str, or pandas series, optional
             Values of the kronecker delta function peak used to encode the event. If a string, the
             column in trialsdf with that label will be used. If a pandas series with indexes
@@ -92,9 +153,7 @@ class DesignMatrixConstructor:
         else:
             raise TypeError('deltaval must be None, pandas series, or string reference'
                             f' to trialsdf column. {type(deltaval)} was passed instead.')
-        if eventname not in self.vartypes:
-            raise ValueError('Event name specified not found in trialsdf')
-        elif self.vartypes[eventname] != 'timing':
+        if eventname in self.vartypes and self.vartypes[eventname] != 'timing':
             raise TypeError(f'Column {eventname} in trialsdf is not registered as a timing')
 
         vecsizes = self.trialsdf['duration'].apply(self.binf)
