+++ brainbox/modeling/utils.py
@@ -1,0 +1,35 @@
+import numpy as np
+
+def raised_cosine(duration, nbases, binfun):
+    nbins = binfun(duration)
+    ttb = repmat(np.arange(1, nbins + 1).reshape(-1, 1), 1, nbases)
+    dbcenter = nbins / nbases
+    cwidth = 4 * dbcenter
+    bcenters = 0.5 * dbcenter + dbcenter * np.arange(0, nbases)
+    x = ttb - repmat(bcenters.reshape(1, -1), nbins, 1)
+    bases = (np.abs(x / cwidth) < 0.5) * (np.cos(x * np.pi * 2 / cwidth) * 0.5 + 0.5)
+    return bases
+
+
+def full_rcos(duration, nbases, binfun, n_before=1):
+    if not isinstance(n_before, int):
+        n_before = int(n_before)
+    nbins = binfun(duration)
+    ttb = repmat(np.arange(1, nbins + 1).reshape(-1, 1), 1, nbases)
+    dbcenter = nbins / (nbases - 2)
+    cwidth = 4 * dbcenter
+    bcenters = 0.5 * dbcenter + dbcenter * np.arange(-n_before, nbases - n_before)
+    x = ttb - repmat(bcenters.reshape(1, -1), nbins, 1)
+    bases = (np.abs(x / cwidth) < 0.5) * (np.cos(x * np.pi * 2 / cwidth) * 0.5 + 0.5)
+    return bases
+
+
+def neglog(weights, x, y):
+    xproj = x @ weights
+    f = np.exp(xproj)
+    nzidx = f != 0
+    if np.any(y[~nzidx] != 0):
+        return np.inf
+    return -y[nzidx].reshape(1, -1) @ xproj[nzidx] + np.sum(f)
+
+
